+++ app/src/main/java/com/example/sstep/date/Date.java
@@ -1,0 +1,15 @@
+package com.example.sstep.date;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.sstep.R;
+public class Date extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_date);
+    }
+}
