+++ app/src/main/java/com/example/sstep/date/Date_modify.java
@@ -1,0 +1,16 @@
+package com.example.sstep.date;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.sstep.R;
+
+public class Date_modify extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_date_modify);
+    }
+}
