+++ app/src/main/java/com/example/sstep/start/Start3.java
@@ -1,4 +1,4 @@
-package com.example.sstep;
+package com.example.sstep.start;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.sstep.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link Start1#newInstance} factory method to
+ * Use the {@link Start3#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Start1 extends Fragment {
+public class Start3 extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class Start1 extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public Start1() {
+    public Start3() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class Start1 extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Start1.
+     * @return A new instance of fragment Start3.
      */
     // TODO: Rename and change types and number of parameters
-    public static Start1 newInstance(String param1, String param2) {
-        Start1 fragment = new Start1();
+    public static Start3 newInstance(String param1, String param2) {
+        Start3 fragment = new Start3();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class Start1 extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_start1, container, false);
+        return inflater.inflate(R.layout.fragment_start3, container, false);
     }
 }
