+++ app/src/main/res/drawable/j_button_date_plus.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <stroke
+        android:width="1dp"
+        android:color="#660D69F3" />
+    <solid android:color="#FFFFFF" />
+    <corners android:radius="100dp" />
+</shape>
