+++ app/src/main/res/drawable/j_button_strokec6c6c6.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#FFFFFF" />
+    <stroke
+        android:width="0.5dp"
+        android:color="#C6C6C6" />
+    <corners android:radius="100dp" />
+</shape>
