+++ app/src/main/res/drawable/j_date_nodataimg.xml
@@ -1,0 +1,78 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="118dp"
+    android:height="55dp"
+    android:viewportWidth="118"
+    android:viewportHeight="55">
+  <path
+      android:pathData="M29.58,0C26.86,0 24.29,0.62 21.99,1.68C21.36,1.58 20.72,1.52 20.08,1.52C17.14,1.52 14.32,2.63 12.25,4.63C10.17,6.62 9,9.32 9,12.13V30.33H9.03C9.32,40.13 17.69,48 28,48C38.31,48 46.68,40.13 46.97,30.33H47V16.68C47,7.47 39.2,0 29.58,0Z"
+      android:fillColor="#292F33"/>
+  <path
+      android:pathData="M21.78,8C21.78,8 21.77,16.09 14,17.05V26.29C14,33.86 20.27,40 28,40C35.73,40 42,33.86 42,26.29V18.64C24.72,18.06 21.78,8 21.78,8Z"
+      android:fillColor="#F7DECE"/>
+  <path
+      android:pathData="M9,43H47V55H9V43Z"
+      android:fillColor="#7450A8"/>
+  <path
+      android:pathData="M29.33,30H26.67C26.49,30 26.32,29.89 26.2,29.71C26.07,29.52 26,29.27 26,29C26,28.73 26.07,28.48 26.2,28.29C26.32,28.11 26.49,28 26.67,28H29.33C29.51,28 29.68,28.11 29.8,28.29C29.93,28.48 30,28.73 30,29C30,29.27 29.93,29.52 29.8,29.71C29.68,29.89 29.51,30 29.33,30Z"
+      android:fillColor="#C1694F"/>
+  <path
+      android:pathData="M21.6,24C21.18,24 20.77,23.86 20.47,23.61C20.17,23.36 20,23.02 20,22.67V21.33C20,20.98 20.17,20.64 20.47,20.39C20.77,20.14 21.18,20 21.6,20C22.02,20 22.43,20.14 22.73,20.39C23.03,20.64 23.2,20.98 23.2,21.33V22.67C23.2,23.02 23.03,23.36 22.73,23.61C22.43,23.86 22.02,24 21.6,24ZM34.4,24C33.98,24 33.57,23.86 33.27,23.61C32.97,23.36 32.8,23.02 32.8,22.67V21.33C32.8,20.98 32.97,20.64 33.27,20.39C33.57,20.14 33.98,20 34.4,20C34.82,20 35.23,20.14 35.53,20.39C35.83,20.64 36,20.98 36,21.33V22.67C36,23.02 35.83,23.36 35.53,23.61C35.23,23.86 34.82,24 34.4,24Z"
+      android:fillColor="#662113"/>
+  <path
+      android:pathData="M5.25,27.2C5.65,26.37 6.43,25.93 7.18,26.01C7.24,25.9 18,31.04 18,31.04L16.06,33.68C15,35.14 13.15,35.43 11.95,34.33L9.39,31.98L5.94,30.03C5.04,29.52 4.73,28.26 5.25,27.2Z"
+      android:fillColor="#EEC2AD"/>
+  <path
+      android:pathData="M0.38,23.37C-0.17,24 -0.12,24.9 0.51,25.37L8.25,29.4L13.88,33.71C14.5,34.19 15.45,34.06 16,33.43L18,31.16L15.75,29.44L10.67,26.65L2.5,23.1C2.5,23.1 0.93,22.75 0.38,23.37Z"
+      android:fillColor="#F7DECE"/>
+  <path
+      android:pathData="M14.55,28.03C14.55,28.03 45.26,41.33 46.06,41.67C47.92,42.49 51,44.29 51,47.32L50.45,54.62C50.45,54.72 50.41,54.82 50.34,54.89C50.26,54.96 50.16,55 50.06,55H44.85C44.77,55 44.7,54.98 44.63,54.93L14.17,34.88C14.12,34.85 14.08,34.8 14.05,34.74C14.02,34.69 14,34.63 14,34.56V28.38C14,28.32 14.02,28.26 14.05,28.2C14.08,28.14 14.12,28.1 14.18,28.06C14.23,28.03 14.3,28.01 14.36,28C14.43,28 14.49,28.01 14.55,28.03Z"
+      android:fillColor="#9268CA"/>
+  <path
+      android:pathData="M28.3,34.46L27.09,34L17,37.02L19.31,38.5L28.3,34.46ZM36.43,37.78L38,38.42L29.92,45L23.96,44.96L36.43,37.78Z"
+      android:fillColor="#7450A8"/>
+  <path
+      android:pathData="M50.75,27.19C50.57,26.81 50.28,26.49 49.94,26.28C49.6,26.07 49.21,25.97 48.82,26.01C48.76,25.9 38,31.04 38,31.04L39.94,33.68C41,35.14 42.85,35.43 44.05,34.33L46.61,31.98L50.06,30.03C50.96,29.52 51.27,28.25 50.75,27.19Z"
+      android:fillColor="#EEC2AD"/>
+  <path
+      android:pathData="M55.62,23.37C56.17,24 56.12,24.9 55.49,25.37L47.75,29.4L42.12,33.71C41.5,34.19 40.55,34.06 40,33.43L38,31.16L40.25,29.44L45.33,26.65L53.5,23.1C53.5,23.1 55.07,22.75 55.62,23.37Z"
+      android:fillColor="#F7DECE"/>
+  <path
+      android:pathData="M41.45,28.03C41.45,28.03 10.74,41.33 9.94,41.67C8.08,42.49 5,44.28 5,47.32L5.55,54.62C5.55,54.83 5.73,55 5.94,55H11.15C11.23,55 11.3,54.98 11.37,54.93L41.83,34.88C41.88,34.85 41.92,34.8 41.95,34.74C41.98,34.69 42,34.63 42,34.56V28.38C42,28.32 41.98,28.26 41.95,28.2C41.92,28.14 41.88,28.1 41.82,28.06C41.77,28.03 41.7,28.01 41.64,28C41.57,28 41.51,28.01 41.45,28.03Z"
+      android:fillColor="#9268CA"/>
+  <path
+      android:pathData="M106,20.02C106,17.63 104.78,15.69 103.26,15.61V8.48L74.74,8V15.61C73.22,15.69 72,17.63 72,20.02C72,22.41 73.22,24.35 74.74,24.43V25.09C74.74,32.77 81.12,39 89,39C96.88,39 103.26,32.77 103.26,25.09V24.42C104.78,24.35 106,22.41 106,20.02Z"
+      android:fillColor="#F7DECE"/>
+  <path
+      android:pathData="M89.5,0C78.08,0 73,7.51 73,13.78C73,20.04 75.19,21.78 75.54,20.04C76.81,13.78 80.62,11.86 80.62,11.86C88.23,18.12 86.96,13.11 86.96,13.11C94.58,20.62 90.77,13.11 90.77,13.11C93.31,15.61 100.92,15.61 100.92,15.61C100.92,15.61 102.19,17.54 103.46,20.04C104.73,22.55 106,20.04 106,13.78C106,7.51 99.65,0 89.5,0Z"
+      android:fillColor="#292F33"/>
+  <path
+      android:pathData="M70,43H108V55H70V43Z"
+      android:fillColor="#2A6797"/>
+  <path
+      android:pathData="M91.17,29H87.83C87.61,29 87.4,28.95 87.24,28.85C87.09,28.76 87,28.63 87,28.5C87,28.37 87.09,28.24 87.24,28.15C87.4,28.05 87.61,28 87.83,28H91.17C91.39,28 91.6,28.05 91.76,28.15C91.91,28.24 92,28.37 92,28.5C92,28.63 91.91,28.76 91.76,28.85C91.6,28.95 91.39,29 91.17,29Z"
+      android:fillColor="#C1694F"/>
+  <path
+      android:pathData="M82.6,24C82.18,24 81.77,23.82 81.47,23.51C81.17,23.2 81,22.78 81,22.33V20.67C81,20.22 81.17,19.8 81.47,19.49C81.77,19.18 82.18,19 82.6,19C83.02,19 83.43,19.18 83.73,19.49C84.03,19.8 84.2,20.22 84.2,20.67V22.33C84.2,22.78 84.03,23.2 83.73,23.51C83.43,23.82 83.02,24 82.6,24ZM95.4,24C94.98,24 94.57,23.82 94.27,23.51C93.97,23.2 93.8,22.78 93.8,22.33V20.67C93.8,20.22 93.97,19.8 94.27,19.49C94.57,19.18 94.98,19 95.4,19C95.82,19 96.23,19.18 96.53,19.49C96.83,19.8 97,20.22 97,20.67V22.33C97,22.78 96.83,23.2 96.53,23.51C96.23,23.82 95.82,24 95.4,24Z"
+      android:fillColor="#662113"/>
+  <path
+      android:pathData="M66.25,27.06C66.65,26.33 67.43,25.94 68.18,26.01C68.24,25.91 79,30.48 79,30.48L77.06,32.83C76,34.13 74.15,34.39 72.95,33.4L70.39,31.31L66.94,29.58C66.04,29.13 65.73,28.01 66.25,27.06Z"
+      android:fillColor="#EEC2AD"/>
+  <path
+      android:pathData="M61.38,22.41C60.83,23.09 60.88,24.07 61.51,24.59L69.25,28.99L74.88,33.68C75.5,34.2 76.45,34.07 77,33.38L79,30.9L76.75,29.02L71.67,25.98L63.5,22.11C63.5,22.11 61.93,21.72 61.38,22.41Z"
+      android:fillColor="#F7DECE"/>
+  <path
+      android:pathData="M75.57,28.03C75.57,28.03 107.1,41.33 107.93,41.67C109.84,42.49 113,44.29 113,47.32L112.43,54.62C112.43,54.72 112.39,54.82 112.32,54.89C112.24,54.96 112.14,55 112.03,55H106.68C106.61,55 106.53,54.98 106.46,54.93L75.18,34.88C75.12,34.85 75.08,34.8 75.05,34.74C75.02,34.69 75,34.63 75,34.56V28.38C75,28.32 75.02,28.26 75.05,28.2C75.08,28.14 75.13,28.1 75.18,28.06C75.24,28.03 75.3,28.01 75.37,28C75.44,28 75.51,28.01 75.57,28.03Z"
+      android:fillColor="#4289C1"/>
+  <path
+      android:pathData="M89.84,33.51L88.57,33L78,36.3L80.42,37.91L89.84,33.51ZM98.35,37.13L100,37.82L91.54,45L85.29,44.96L98.35,37.13Z"
+      android:fillColor="#2A6797"/>
+  <path
+      android:pathData="M112.75,27.06C112.57,26.72 112.29,26.43 111.94,26.25C111.6,26.06 111.21,25.98 110.82,26.01C110.76,25.91 100,30.48 100,30.48L101.94,32.83C103,34.13 104.85,34.39 106.05,33.4L108.61,31.31L112.06,29.58C112.96,29.13 113.26,28 112.75,27.06Z"
+      android:fillColor="#EEC2AD"/>
+  <path
+      android:pathData="M117.62,22.41C118.17,23.09 118.12,24.07 117.49,24.59L109.75,28.99L104.12,33.68C103.5,34.2 102.55,34.07 102,33.38L100,30.9L102.25,29.02L107.33,25.98L115.5,22.11C115.5,22.11 117.07,21.72 117.62,22.41Z"
+      android:fillColor="#F7DECE"/>
+  <path
+      android:pathData="M102.45,28.03C102.45,28.03 71.74,41.33 70.94,41.67C69.08,42.49 66,44.28 66,47.32L66.55,54.62C66.55,54.83 66.73,55 66.94,55H72.15C72.23,55 72.3,54.98 72.37,54.93L102.83,34.88C102.88,34.85 102.92,34.8 102.95,34.74C102.98,34.69 103,34.63 103,34.56V28.38C103,28.32 102.98,28.26 102.95,28.2C102.92,28.14 102.88,28.1 102.82,28.06C102.77,28.03 102.71,28.01 102.64,28C102.57,28 102.51,28.01 102.45,28.03Z"
+      android:fillColor="#4289C1"/>
+</vector>
