+++ app/src/main/res/drawable/j_repeat_radio_btn.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="false"
+        android:drawable="@drawable/j_button_strokec6c6c6" />
+    <item android:state_checked="true"
+        android:drawable="@drawable/j_button_stroke070707"/>
+
+
+
+</selector>
