+++ app/src/main/res/drawable/j_repeat_radio_repeatimg.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="false"
+        android:drawable="@drawable/j_repeat_unchecked" />
+    <item android:state_checked="true"
+        android:drawable="@drawable/j_repeat"/>
+
+
+
+</selector>
