+++ app/src/main/res/drawable/j_repeat_radio_text.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="#D9000000">
+    </item>
+    <item android:state_checked="false" android:color="#4D000000">
+    </item>
+</selector>
