+++ app/src/main/res/drawable/j_repeat_radio_ximg.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="false"
+        android:drawable="@drawable/j_lightblack_x_img" />
+    <item android:state_checked="true"
+        android:drawable="@drawable/j_black_x_img"/>
+
+
+
+</selector>
