+++ app/src/main/res/drawable/j_spinner_backgroundcdcdc35.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#26DCDCDC" />
+    <corners android:radius="10dp" />
+</shape>
